+++ packages/framework/src/Foundation/Kernel/Hyperlinks.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Hyde\Foundation\Kernel;
 
 use Hyde\Facades\Config;
+use BadMethodCallException;
 use Hyde\Support\Models\Route;
 use Hyde\Foundation\HydeKernel;
 use Hyde\Framework\Exceptions\FileNotFoundException;
@@ -110,7 +111,7 @@ class Hyperlinks
      */
     public function asset(string $name, bool $preferQualifiedUrl = false): string
     {
-        if (str_starts_with($name, 'http')) {
+        if (static::isRemote($name)) {
             return $name;
         }
 
@@ -139,13 +140,15 @@ class Hyperlinks
      * Return a qualified URL to the supplied path if a base URL is set.
      *
      * @param  string  $path  An optional relative path suffix. Omit to return the base URL.
-     * @return string|null The qualified URL, or null if the base URL is not set and no path is provided.
+     * @return string The qualified URL, or the base URL if no path is supplied.
+     *
+     * @throws BadMethodCallException If the site URL is not set in the configuration and no path is supplied.
      */
-    public function url(string $path = ''): ?string
+    public function url(string $path = ''): string
     {
         $path = $this->formatLink(trim($path, '/'));
 
-        if (str_starts_with($path, 'http')) {
+        if (static::isRemote($path)) {
             return $path;
         }
 
@@ -159,8 +162,8 @@ class Hyperlinks
             return $path;
         }
 
-        // User is trying to get the base URL, but it's not set
-        return null;
+        // User is trying to get the base URL, but it's not set, so we throw an exception.
+        throw new BadMethodCallException('The site URL is not set in the configuration.');
     }
 
     /**
@@ -169,5 +172,13 @@ class Hyperlinks
     public function route(string $key): ?Route
     {
         return $this->kernel->routes()->get($key);
+    }
+
+    /**
+     * Determine if the given URL is a remote link.
+     */
+    public static function isRemote(string $url): bool
+    {
+        return str_starts_with($url, 'http') || str_starts_with($url, '//');
     }
 }