+++ packages/framework/src/Framework/Features/Blogging/Models/FeaturedImage.php
@@ -10,6 +10,7 @@ use Hyde\Facades\Config;
 use Illuminate\Support\Str;
 use Hyde\Support\BuildWarnings;
 use Illuminate\Support\Facades\Http;
+use Hyde\Foundation\Kernel\Hyperlinks;
 use Hyde\Framework\Exceptions\FileNotFoundException;
 use Hyde\Markdown\Contracts\FrontMatter\SubSchemas\FeaturedImageSchema;
 
@@ -19,7 +20,6 @@ use function file_exists;
 use function filesize;
 use function sprintf;
 use function key;
-use function str_starts_with;
 
 /**
  * Object representation of a blog post's featured image.
@@ -63,7 +63,7 @@ class FeaturedImage implements Stringable, FeaturedImageSchema
         protected readonly ?string $licenseUrl = null,
         protected readonly ?string $copyrightText = null
     ) {
-        $this->type = self::isRemote($source) ? self::TYPE_REMOTE : self::TYPE_LOCAL;
+        $this->type = Hyperlinks::isRemote($source) ? self::TYPE_REMOTE : self::TYPE_LOCAL;
         $this->source = $this->setSource($source);
     }
 
@@ -239,10 +239,5 @@ class FeaturedImage implements Stringable, FeaturedImageSchema
         }
 
         return 0;
-    }
-
-    public static function isRemote(string $source): bool
-    {
-        return str_starts_with($source, 'http') || str_starts_with($source, '//');
     }
 }